Redis pubsub, organise header to avoid exceptions while waiting for signup
