+++ package.json
@@ -42,6 +42,7 @@
     "@radix-ui/react-toggle": "1.0.3",
     "@radix-ui/react-toggle-group": "1.0.4",
     "@radix-ui/react-tooltip": "1.0.7",
+    "@redis/client": "^1.6.0",
     "@sentry/nextjs": "8.20.0",
     "@square/web-sdk": "2.0.1",
     "@t3-oss/env-nextjs": "0.9.2",
