+++ src/server/api/ratelimit.ts
@@ -1,19 +1,11 @@
 import { type ProcedureType } from "@trpc/server"
 import { Ratelimit, type RatelimitConfig } from "@upstash/ratelimit"
-import { Redis } from "@upstash/redis"
+import { createRedis } from "./redis"
 import { type TRPCContext } from "./trpc"
-import { env } from "~/env"
 
 export const createRatelimit = (limiter: RatelimitConfig["limiter"]) =>
   new Ratelimit({
-    redis: new Redis({
-      url: env.UPSTASH_REDIS_REST_URL,
-      token: env.UPSTASH_REDIS_REST_TOKEN,
-      retry: {
-        retries: 5,
-        backoff: (retryCount) => Math.exp(retryCount) * 100,
-      },
-    }),
+    redis: createRedis(),
     limiter,
     ephemeralCache: new Map(),
     analytics: true,
